add in support for plenary service templates (under the "servicedata" template namespace, since we don't want to clash with the "service" user-maintained area). The only data put into the plenary templates (for now) is the instance name and the servers which are bound to the instance. To test the new stuff, I've also added a "flush" command which writes out *all* plenary templates (in fact, this will be neccessary during migration, in order to get the new templates written out, instead of waiting for them to be modified by users). For now, the plenary templates are NOT included from the standard templates (because that would break anyone's templates out there right now), and the template changes will happen via a sandbox deploy after 1.2.1 has made it out. I've not made unittests yet, since I don't quite grok the system. That'll be next...
[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 647465]

+++ 1.2.1/src/lib/python2.5/aquilon/server/commands/del_service.py
@@ -18,6 +18,8 @@ from aquilon.server.broker import (format_results, add_transaction, az_check,
 from aquilon.server.dbwrappers.service import get_service
 from aquilon.aqdb.svc.service_instance import ServiceInstance
 from aquilon.aqdb.sy.host_list import HostList
+from aquilon.aqdb.svc.service_map import ServiceMap
+from aquilon.server.templates import (PlenaryService, PlenaryServiceInstance)
 
 
 class CommandDelService(BrokerCommand):
@@ -32,6 +34,8 @@ class CommandDelService(BrokerCommand):
         if not instance:
             if dbservice.instances:
                 raise ArgumentError("Cannot remove service with instances defined.")
+            plenary_info = PlenaryService(dbservice)
+            plenary_info.remove
             session.delete(dbservice)
             return
         try:
@@ -42,9 +46,21 @@ class CommandDelService(BrokerCommand):
                     % (instance, e))
         dbsi = session.query(ServiceInstance).filter_by(
                 host_list=dbhl, service=dbservice).first()
-        # FIXME: There may be dependencies...
+
         if dbsi:
+            if dbsi.client_count > 0:
+                raise ArgumentError("instance has clients and cannot be deleted.")
+
+            # Check the service map and remove any mappings
+            for dbmap in session.query(ServiceMap).filter_by(service_instance=dbsi).all():
+                session.delete(dbmap)
+
+            plenary_info = PlenaryServiceInstance(dbservice, dbsi)
+            plenary_info.remove
             session.delete(dbsi)
+            session.flush()
+            session.refresh(dbservice)
+            
         # FIXME: Cascade to relevant objects...
         return
